Fix checksum calculation (#215)

+++ src/main/java/ru/yandex/clickhouse/util/ClickHouseCityHash.java
@@ -59,15 +59,15 @@ public class ClickHouseCityHash {
                 ((b[i+1] & 255) <<  8) +
                 ((b[i+0] & 255) <<  0));
     }
-    private  static int toIntLE(byte[] b, int i) {
-        return (((b[i+3] & 255) << 24) + ((b[i+2] & 255) << 16) + ((b[i+1] & 255) << 8) + ((b[i+0] & 255) << 0));
+    private static long toIntLE(byte[] b, int i) {
+        return (((b[i+3] & 255L) << 24) + ((b[i+2] & 255L) << 16) + ((b[i+1] & 255L) << 8) + ((b[i+0] & 255L) << 0));
     }
 
     private static long fetch64(byte[] s, int pos) {
         return toLongLE(s, pos);
     }
 
-    private static int fetch32(byte[] s, int pos) {
+    private static long fetch32(byte[] s, int pos) {
         return toIntLE(s, pos);
     }
